# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kalapi Restaurant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface SignupRequest {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  user: "user",
  admin: "admin",
} as const;

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  role: UserProfileRole;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface UpdateProfileRequest {
  name?: string;
  phone?: string;
}

export interface MenuItem {
  id: number;
  name: string;
  description?: string;
  price: number;
  category: string;
  image?: string;
  isVeg: boolean;
  isAvailable: boolean;
  isFeatured?: boolean;
  rating?: number;
  createdAt?: string;
}

export interface CreateMenuItemRequest {
  name: string;
  description?: string;
  price: number;
  category: string;
  image?: string;
  isVeg: boolean;
  isAvailable?: boolean;
  isFeatured?: boolean;
}

export interface CartItem {
  id: number;
  menuItemId: number;
  quantity: number;
  menuItem?: MenuItem;
}

export interface AddToCartRequest {
  menuItemId: number;
  quantity: number;
}

export interface UpdateCartItemRequest {
  quantity: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  preparing: "preparing",
  out_for_delivery: "out_for_delivery",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type OrderPaymentStatus =
  (typeof OrderPaymentStatus)[keyof typeof OrderPaymentStatus];

export const OrderPaymentStatus = {
  unpaid: "unpaid",
  paid: "paid",
  failed: "failed",
} as const;

export interface OrderItem {
  id: number;
  menuItemId: number;
  quantity: number;
  price: number;
  menuItem?: MenuItem;
}

export interface Order {
  id: number;
  userId: number;
  items?: OrderItem[];
  total: number;
  status: OrderStatus;
  paymentStatus: OrderPaymentStatus;
  paymentId?: string;
  deliveryAddress?: string;
  deliveryName?: string;
  deliveryPhone?: string;
  couponCode?: string;
  discount?: number;
  createdAt?: string;
}

export interface PlaceOrderRequest {
  deliveryName: string;
  deliveryPhone: string;
  deliveryAddress: string;
  deliveryPincode?: string;
  couponCode?: string;
}

export interface PaymentRequest {
  razorpayPaymentId: string;
  razorpayOrderId?: string;
  razorpaySignature?: string;
}

export type UpdateOrderStatusRequestStatus =
  (typeof UpdateOrderStatusRequestStatus)[keyof typeof UpdateOrderStatusRequestStatus];

export const UpdateOrderStatusRequestStatus = {
  pending: "pending",
  preparing: "preparing",
  out_for_delivery: "out_for_delivery",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusRequest {
  status: UpdateOrderStatusRequestStatus;
}

export interface Address {
  id: number;
  label?: string;
  street: string;
  city: string;
  pincode: string;
  isDefault?: boolean;
}

export interface AddressRequest {
  label?: string;
  street: string;
  city: string;
  pincode: string;
  isDefault?: boolean;
}

export interface WishlistRequest {
  menuItemId: number;
}

export interface WishlistResponse {
  added: boolean;
  menuItemId: number;
}

export interface ContactRequest {
  name: string;
  email: string;
  phone?: string;
  message: string;
  bookingDate?: string;
}

export type GetMenuParams = {
  category?: string;
  search?: string;
};