# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kalapi Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderPaymentStatus } from "./orderPaymentStatus";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  userId: number;
  items?: OrderItem[];
  total: number;
  status: OrderStatus;
  paymentStatus: OrderPaymentStatus;
  paymentId?: string;
  deliveryAddress?: string;
  deliveryName?: string;
  deliveryPhone?: string;
  couponCode?: string;
  discount?: number;
  createdAt?: Date;
}