# lib/api-zod/src/generated/types/paymentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kalapi Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentRequest {
  razorpayPaymentId: string;
  razorpayOrderId?: string;
  razorpaySignature?: string;
}